==> docs/00_project_setup.rmd <==
---
title: "00 Project Setup"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>

This file described dependencies and potential troubleshooting tips.

This paper was run using v1.0.0 of the MeasuringLandscape package. This release reflects the state of the code as of publication. It is the best chance at obtaining identical results as shown in print. Future releases will improve the presentation of the R-Notebooks and refactor the code to be more readable and reusable and may not generate exactly identical figures or messages. This version is archived so that even as the package improves over time, the analysis can be rerun against that specific version of the code.

This package is highly dependent on specific package versions. Many packages updated their interfaces just over the course of writing this paper. These are the specific versions of packages used. 

```{r}
sessionInfo()

# install.packages("devtools") 
#devtools::install_github("rexdouglass/MeasuringLandscape") #@v1.0.0
```

You may hit an error "`maximal number of DLLs reached..." This is set by an environmental variable in R. We've specifcied it in the included ".Renviron" file "R_MAX_NUM_DLLS = 614"" which should work unless your system has a hard fixed lower level or you specifcy it elsehwere.

MeasuringLandscape depends on a number of other packages, some with their own complicated dependencies.

First SF, which requires GDAL. Follow the [install instructions for SF](https://github.com/r-spatial/sf) which vary depending on your operating system.

``` r
install.packages('sf') #will either work or fail depending on when you have GDAL installed
```

Next we recomend all of tidyverse and tidymodels

``` r
install.packages("tidyverse")
```


In order to use specific fonts in the pdf, we use a system called Cairo. To use Cairo, must also install from https://www.cairographics.org/download/
```{r, eval=F}
#configure: error: Cannot find cairo.h! Please install cairo (http://www.cairographics.org/) and/or set CAIRO_CFLAGS/LIBS correspondingly
#sudo apt install r-cran-cairo
install.packages('Cairo', dependencies=T)
```


```{r, eval=F}
if(!require(devtools)) install.packages("devtools")
devtools::install_github("dselivanov/LSHR") 
```

```{r}
install.packages('data.table', dependencies=T)
```

```{r}
install.packages('sf', dependencies=T)
```

```{r}
install.packages('digest', dependencies=T)
```


```{r}
install.packages('janitor', dependencies=T)
```

```{r}
install.packages('mosaic', dependencies=T)
install.packages('janitor', dependencies=T)
```

```{r}
install.packages('here')
```

```{r}
install.packages('re2r', dependencies=T)
```

```{r}
install.packages('xgboost', dependencies=T)
```

```{r}
install.packages('dummies', dependencies=T)
install.packages('fastDummies', dependencies=T)

```

```{r}
install.packages('doParallel', dependencies=T)
install.packages('quanteda', dependencies=T)
```

```{r}
install.packages('rgdal', dependencies=T)
install.packages('raster', dependencies=T)
```

```{r}
install.packages('Metrics', dependencies=T)
install.packages('MLmetrics', dependencies=T)
install.packages('strigr', dependencies=T)
install.packages('cowplot', dependencies=T)
install.packages('units', dependencies=T)
install.packages('ggrepel', dependencies=T)
install.packages('ggmap', dependencies=T)
install.packages('ggformula', dependencies=T)
install.packages('precrec', dependencies=T)

```

```{r}
# must download 

devtools::document()

#library("miniCRAN")
#tags <- "MeasuringLandscape"
#pkgDep(tags, availPkgs = cranJuly2014)

require(devtools)
#install_version("backports", version = "1.1.0")
#library(backports)

```
